+++ hazelcast/src/main/java/com/hazelcast/eviction/EvictionConfiguration.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache.impl.eviction;
+package com.hazelcast.eviction;
 
 /**
  * Interface for configuration information about eviction.
