+++ hazelcast/src/main/java/com/hazelcast/eviction/EvictionPolicyEvaluator.java
@@ -14,10 +14,14 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache.impl.eviction;
+package com.hazelcast.eviction;
 
 /**
  * Interface for evaluation implementations of {@link com.hazelcast.config.EvictionPolicy}.
+ *
+ * @param <A> Type of the accessor (id) of the {@link com.hazelcast.eviction.EvictionCandidate}
+ * @param <E> Type of the {@link com.hazelcast.eviction.Evictable} value of
+ *            {@link com.hazelcast.eviction.EvictionCandidate}
  */
 public interface EvictionPolicyEvaluator<A, E extends Evictable> {
