+++ hazelcast/src/main/java/com/hazelcast/eviction/EvictionPolicyEvaluatorProvider.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache.impl.eviction;
+package com.hazelcast.eviction;
 
-import com.hazelcast.cache.impl.eviction.impl.evaluator.LFUEvictionPolicyEvaluator;
-import com.hazelcast.cache.impl.eviction.impl.evaluator.LRUEvictionPolicyEvaluator;
+import com.hazelcast.eviction.impl.evaluator.LFUEvictionPolicyEvaluator;
+import com.hazelcast.eviction.impl.evaluator.LRUEvictionPolicyEvaluator;
 
 import java.util.HashMap;
 import java.util.Map;
