+++ hazelcast/src/main/java/com/hazelcast/eviction/Expirable.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache.impl.eviction;
+package com.hazelcast.eviction;
 
 /**
  * Expiring Data model interface.
