+++ hazelcast/src/main/java/com/hazelcast/eviction/impl/package-info.java
@@ -14,21 +14,9 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cache.impl.eviction;
-
 /**
- * Interface for entries, records or whatever that can be checked for expiration.
- *
- * @param <E> Type of the {@link Expirable} value
+ * <p>
+ *     Eviction implementations for cache.
+ * </p>
  */
-public interface ExpirationChecker<E extends Expirable> {
-
-    /**
-     * Checks if the given {@link Expirable} entry is expired or not.
-     *
-     * @param expirableEntry {@link Expirable} entry which is checked to see if it is expired or not.
-     * @return <code>true</code> if the {@link Expirable} entry is expired, otherwise <code>false</code>.
-     */
-    boolean isExpired(E expirableEntry);
-
-}
+package com.hazelcast.eviction.impl;
