+++ hazelcast/src/main/java/com/hazelcast/util/CacheConcurrentHashMap.java
@@ -17,7 +17,7 @@
 package com.hazelcast.util;
 
 import com.hazelcast.cache.impl.CacheKeyIteratorResult;
-import com.hazelcast.cache.impl.eviction.Expirable;
+import com.hazelcast.eviction.Expirable;
 import com.hazelcast.nio.serialization.Data;
 
 import java.util.ArrayList;
